move libass functionality to subclass
