+++ include/mpc-hc_config.h
@@ -40,7 +40,7 @@
 
 #define USE_STATIC_UNRAR 1
 
-#define USE_LIBASS 0
+#define USE_LIBASS 1
 
 #ifndef USE_DRDUMP_CRASH_REPORTER
 #define USE_DRDUMP_CRASH_REPORTER 0
