+++ src/Subtitles/RTS.cpp
@@ -22,11 +22,9 @@
 #include "stdafx.h"
 #include <cmath>
 #include <intrin.h>
-#include <algorithm>
 #include "ColorConvTable.h"
 #include "RTS.h"
 #include "../DSUtil/PathUtils.h"
-#include <ppl.h>
 #include "../filters/renderer/VideoRenderers/RenderersSettings.h"
 
 // WARNING: this isn't very thread safe, use only one RTS a time. We should use TLS in future.
@@ -3119,54 +3117,6 @@ struct LSub {
     }
 };
 
-namespace {
-    inline POINT GetRectPos(RECT rect) {
-        return { rect.left, rect.top };
-    }
-
-    inline SIZE GetRectSize(RECT rect) {
-        return { rect.right - rect.left, rect.bottom - rect.top };
-    }
-}
-
-#if USE_LIBASS
-void AssFlatten(ASS_Image* image, SubPicDesc& spd, CRect &rcDirty) {
-    if (image) {
-        RECT pRect = { 0 };
-        for (auto i = image; i != nullptr; i = i->next) {
-            RECT rect1 = pRect;
-            RECT rect2 = { i->dst_x, i->dst_y, i->dst_x + i->w, i->dst_y + i->h };
-            UnionRect(&pRect, &rect1, &rect2);
-        }
-
-        const POINT pixelsPoint = GetRectPos(pRect);
-        const SIZE pixelsSize = GetRectSize(pRect);
-        rcDirty.IntersectRect(CRect(pixelsPoint, pixelsSize), CRect(0, 0, spd.w, spd.h));
-
-        BYTE* pixelBytes = (BYTE*)(spd.bits + spd.pitch * rcDirty.top + rcDirty.left * 4);
-
-        for (auto i = image; i != nullptr; i = i->next) {
-            concurrency::parallel_for(0, i->h, [&](int y)
-                {
-                    for (int x = 0; x < i->w; ++x) {
-                        BYTE* dst = &pixelBytes[((ptrdiff_t)i->dst_y + y - pixelsPoint.y) * spd.pitch + ((ptrdiff_t)i->dst_x + x - pixelsPoint.x) * 4];
-
-                        uint32_t srcA = (i->bitmap[y * i->stride + x] * (0xff - (i->color & 0x000000ff))) >> 8;
-                        uint32_t compA = 0xff - srcA;
-
-
-                        dst[3] = 0xff - (srcA + (((0xff-dst[3]) * compA) >> 8)); //A.  this is inverted alpha, so we invert it before multiplying and then invert it again
-                        dst[2] = (((i->color & 0xff000000) >> 24) * srcA + (dst[2]) * compA) >> 8; //R
-                        dst[1] = (((i->color & 0x00ff0000) >> 16) * srcA + dst[1] * compA) >> 8; //G
-                        dst[0] = (((i->color & 0x0000ff00) >> 8) * srcA + dst[0] * compA) >> 8; //B
-
-                    }
-                }, concurrency::static_partitioner());
-        }
-    }
-}
-#endif
-
 STDMETHODIMP CRenderedTextSubtitle::Render(SubPicDesc& spd, REFERENCE_TIME rt, double fps, RECT& bbox)
 {
     CAutoLock cAutoLock(&renderLock);
@@ -3177,35 +3127,27 @@ STDMETHODIMP CRenderedTextSubtitle::Render(SubPicDesc& spd, REFERENCE_TIME rt, d
         return S_FALSE;
     }
 
-#if USE_LIBASS
-    if (m_assloaded) {
+    if (m_SSAUtil.m_assloaded) {
         if (spd.bpp != 32) {
             ASSERT(FALSE);
             return E_INVALIDARG;
         }
 
-        if (!m_assfontloaded && m_pPin) {
-            LoadASSFont(m_pPin, m_ass.get(), m_renderer.get());
-        }
+        m_SSAUtil.LoadASSFont();
 
         m_size = CSize(spd.w, spd.h);
         m_vidrect = CRect(spd.vidrect.left, spd.vidrect.top, spd.vidrect.right, spd.vidrect.bottom);
-        ass_set_frame_size(m_renderer.get(), spd.w, spd.h);
-
-        int changed = 1;
-        ASS_Image* image = ass_render_frame(m_renderer.get(), m_track.get(), rt / 10000, &changed);
-
-        if (!image) {
-            return E_FAIL;
-        }
+        m_SSAUtil.SetFrameSize(spd.w, spd.h);
 
         CRect rcDirty;
-        AssFlatten(image, spd, rcDirty);
+
+        if (!m_SSAUtil.RenderFrame(rt / 10000, spd, rcDirty)) {
+            return E_FAIL;
+        }
 
         bbox = rcDirty;
         return S_OK;
     }
-#endif
 
     Init(CSize(spd.w, spd.h), spd.vidrect);
