+++ src/Subtitles/STS.cpp
@@ -2308,25 +2308,13 @@ CSimpleTextSubtitle::CSimpleTextSubtitle()
     , m_bUsingPlayerDefaultStyle(false)
     , m_ePARCompensationType(EPCTDisabled)
     , m_dPARCompensation(1.0)
-#if USE_LIBASS
-    , m_renderUsingLibass(true)
-    , m_openTypeLangHint()
-    , m_assloaded(false)
-    , m_assfontloaded(false)
-    , m_pGraph(nullptr)
-    , m_pPin(nullptr)
-    , m_ass(nullptr)
-    , m_renderer(nullptr)
-    , m_track(nullptr)
-#endif
+    , m_SSAUtil(this)
 {
 }
 
 CSimpleTextSubtitle::~CSimpleTextSubtitle()
 {
-#if USE_LIBASS
-    UnloadASS();
-#endif
+    m_SSAUtil.UnloadASS();
 
     Empty();
 }
@@ -2367,11 +2355,9 @@ void CSimpleTextSubtitle::Copy(CSimpleTextSubtitle& sts)
         m_segments.Copy(sts.m_segments);
         __super::Copy(sts);
 
-#if USE_LIBASS
-        if (m_assloaded) {
-            LoadASSFile(m_subtitleType);
+        if (m_SSAUtil.m_assloaded) {
+            m_SSAUtil.LoadASSFile(m_subtitleType);
         }
-#endif
     }
 }
 
@@ -3217,12 +3203,11 @@ bool CSimpleTextSubtitle::Open(CTextFile* f, int CharSet, CString name) {
         m_langname = ISOLang::LCIDToLanguage(m_lcid);
     }
 
-#if USE_LIBASS
-    if (m_renderUsingLibass) {
+    if (m_SSAUtil.m_renderUsingLibass) {
         if (lstrcmpi(PathFindExtensionW(f->GetFilePath()), L".ass") == 0 || lstrcmpi(PathFindExtensionW(f->GetFilePath()), L".ssa") == 0) {
             CreateDefaultStyle(CharSet);
             m_path = f->GetFilePath();
-            LoadASSFile(Subtitle::SubType::SSA);
+            m_SSAUtil.LoadASSFile(Subtitle::SubType::SSA);
             m_subtitleType = Subtitle::SubType::SSA;
             OpenSubStationAlpha(f, *this, CharSet);
             CWebTextFile f2(CTextFile::UTF8);
@@ -3232,12 +3217,12 @@ bool CSimpleTextSubtitle::Open(CTextFile* f, int CharSet, CString name) {
         } else if (lstrcmpi(PathFindExtensionW(f->GetFilePath()), L".srt") == 0) {
             CreateDefaultStyle(CharSet);
             m_path = f->GetFilePath();
-            LoadASSFile(Subtitle::SubType::SRT);
+            m_SSAUtil.LoadASSFile(Subtitle::SubType::SRT);
             m_subtitleType = Subtitle::SubType::SRT;
             OpenSubRipper(f, *this, CharSet);
         }
 
-        if (m_assloaded) {
+        if (m_SSAUtil.m_assloaded) {
             m_name = name;
             m_encoding = f->GetEncoding();
             m_mode = TIME;
@@ -3259,7 +3244,6 @@ bool CSimpleTextSubtitle::Open(CTextFile* f, int CharSet, CString name) {
             return true;
         }
     }
-#endif
 
     ULONGLONG pos = f->GetPosition();
 
@@ -3318,185 +3302,6 @@ bool CSimpleTextSubtitle::Open(CTextFile* f, int CharSet, CString name) {
     m_path = _T("");
     return false;
 }
-
-#if USE_LIBASS
-
-void CSimpleTextSubtitle::SetSubRenderSettings(SubRendererSettings settings) {
-    bool wasUsingLibass = subRendererSettings.renderUsingLibass;
-    subRendererSettings = settings;
-    if (settings.renderUsingLibass || wasUsingLibass) {
-        ResetASS();
-    }
-}
-
-void CSimpleTextSubtitle::ResetASS() {
-    if (subRendererSettings.renderUsingLibass) { 
-        m_renderUsingLibass = true;
-        if (!m_path.IsEmpty()) {
-            LoadASSFile(m_subtitleType);
-        }
-    } else {
-        if (m_assloaded) {
-            UnloadASS();
-        }
-        m_renderUsingLibass = false;
-    }
-}
-
-bool CSimpleTextSubtitle::LoadASSFile(Subtitle::SubType subType) {
-    if (m_path.IsEmpty() || !PathUtils::Exists(m_path)) return false;
-    UnloadASS();
-
-    m_assfontloaded = false;
-
-    m_ass = decltype(m_ass)(ass_library_init());
-    ass_set_extract_fonts(m_ass.get(), true);
-    ass_set_style_overrides(m_ass.get(), NULL);
-
-    m_renderer = decltype(m_renderer)(ass_renderer_init(m_ass.get()));
-    ass_set_use_margins(m_renderer.get(), false);
-    ass_set_font_scale(m_renderer.get(), 1.0);
-
-    STSStyle defStyle;
-    GetDefaultStyle(defStyle);
-    if (subType == Subtitle::SRT) {
-        m_track = decltype(m_track)(srt_read_file(m_ass.get(), const_cast<char*>((const char*)(CStringA)m_path), defStyle.charSet, defStyle, subRendererSettings));
-        if (m_storageRes == CSize(0, 0)) {
-            m_storageRes = CSize(defStyle.SrtResX, defStyle.SrtResY);
-        }
-    } else { //subType == Subtitle::SSA/ASS
-        m_track = decltype(m_track)(ass_read_file(m_ass.get(), const_cast<char*>((const char*)(CStringA)m_path), "UTF-8"));
-        if (m_storageRes == CSize(0, 0)) {
-            m_storageRes = CSize(defStyle.SrtResX, defStyle.SrtResY);
-        }
-    }
-
-    if (!m_track) return false;
-
-    CT2CA tmpFontName(defStyle.fontName);
-    ass_set_fonts(m_renderer.get(), NULL, std::string(tmpFontName).c_str(), ASS_FONTPROVIDER_DIRECTWRITE, NULL, 0);
-
-    m_assloaded = true;
-    m_assfontloaded = true;
-
-    return true;
-}
-
-bool CSimpleTextSubtitle::LoadASSTrack(char* data, int size, Subtitle::SubType subType) {
-    UnloadASS();
-    m_assfontloaded = false;
-
-    m_ass = decltype(m_ass)(ass_library_init());
-    m_renderer = decltype(m_renderer)(ass_renderer_init(m_ass.get()));
-    m_track = decltype(m_track)(ass_new_track(m_ass.get()));
-
-    if (!m_track) return false;
-
-    STSStyle defStyle;
-    GetDefaultStyle(defStyle);
-    if (subType == Subtitle::SRT) {
-        std::stringstream srtData;
-        srtData.write(data, size);
-        srt_read_data(m_ass.get(), m_track.get(), srtData, defStyle.charSet, defStyle, subRendererSettings);
-    } else { //subType == Subtitle::SSA/ASS
-        ass_process_codec_private(m_track.get(), data, size);
-    }
-    CT2CA tmpFontName(defStyle.fontName);
-    ass_set_fonts(m_renderer.get(), NULL, std::string(tmpFontName).c_str(), ASS_FONTPROVIDER_DIRECTWRITE, NULL, 0);
-    //don't set m_assfontloaded here, in case we can load embedded fonts later?
-
-    m_assloaded = true;
-    return true;
-}
-
-void CSimpleTextSubtitle::LoadASSFont(IPin* pPin, ASS_Library* ass, ASS_Renderer* renderer) {
-    // Try to load fonts in the container
-    CComPtr<IAMGraphStreams> graphStreams;
-    CComPtr<IDSMResourceBag> bag;
-    if (m_pGraph && SUCCEEDED(m_pGraph->QueryInterface(IID_PPV_ARGS(&graphStreams))) &&
-        SUCCEEDED(graphStreams->FindUpstreamInterface(pPin, IID_PPV_ARGS(&bag), AM_INTF_SEARCH_FILTER))) {
-        for (DWORD i = 0; i < bag->ResGetCount(); ++i) {
-            _bstr_t name, desc, mime;
-            BYTE* pData = nullptr;
-            DWORD len = 0;
-            if (SUCCEEDED(bag->ResGet(i, &name.GetBSTR(), &desc.GetBSTR(), &mime.GetBSTR(), &pData, &len, nullptr))) {
-                if (wcscmp(mime.GetBSTR(), L"application/x-truetype-font") == 0 ||
-                    wcscmp(mime.GetBSTR(), L"application/vnd.ms-opentype") == 0) // TODO: more mimes?
-                {
-                    ass_add_font(ass, (char*)name, (char*)pData, len);
-                    // TODO: clear these fonts somewhere?
-                }
-                CoTaskMemFree(pData);
-            }
-        }
-        m_assfontloaded = true;
-        STSStyle defStyle;
-        GetDefaultStyle(defStyle);
-        CT2CA tmpFontName(defStyle.fontName);
-        ass_set_fonts(renderer, NULL, std::string(tmpFontName).c_str(), ASS_FONTPROVIDER_DIRECTWRITE, NULL, 0);
-    }
-}
-
-void CSimpleTextSubtitle::UnloadASS() {
-    m_assloaded = false;
-    if (m_track) m_track.reset();
-    if (m_renderer) m_renderer.reset();
-    if (m_ass) m_ass.reset();
-}
-
-void CSimpleTextSubtitle::LoadASSSample(char *data, int dataSize, REFERENCE_TIME tStart, REFERENCE_TIME tStop) {
-    if (m_renderUsingLibass) {
-        if (m_subtitleType == Subtitle::SRT) { //received SRT sample, try to use libass to handle
-            if (!m_assloaded) { //create ass header
-                UnloadASS();
-                m_assfontloaded = false;
-
-                m_ass = decltype(m_ass)(ass_library_init());
-                m_renderer = decltype(m_renderer)(ass_renderer_init(m_ass.get()));
-                m_track = decltype(m_track)(ass_new_track(m_ass.get()));
-
-                char outBuffer[1024];
-                STSStyle defStyle;
-                GetDefaultStyle(defStyle);
-                srt_header(outBuffer, defStyle, subRendererSettings);
-                ass_process_codec_private(m_track.get(), outBuffer, static_cast<int>(strnlen_s(outBuffer, sizeof(outBuffer))));
-                m_assloaded = true;
-            }
-
-            if (m_assloaded) {
-                char subLineData[1024]{};
-                strncpy_s(subLineData, _countof(subLineData), data, dataSize);
-                std::string str = subLineData;
-
-                // This is the way i use to get a unique id for the subtitle line
-                // It will only fail in the case there is 2 or more lines with the same start timecode
-                // (Need to check if the matroska muxer join lines in such a case)
-                REFERENCE_TIME m_iSubLineCount = tStart / 10000;
-
-                // Change srt tags to ass tags
-                STSStyle defStyle;
-                GetDefaultStyle(defStyle);
-                ParseSrtLine(str, defStyle);
-
-                // Add the custom tags
-                CT2CA tmpCustomTags(defStyle.customTags);
-                str.insert(0, std::string(tmpCustomTags));
-
-                // Add blur
-                char blur[20]{};
-                _snprintf_s(blur, _TRUNCATE, "{\\blur%u}", defStyle.fBlur);
-                str.insert(0, blur);
-
-                // ASS in MKV: ReadOrder, Layer, Style, Name, MarginL, MarginR, MarginV, Effect, Text
-                char outBuffer[1024]{};
-                _snprintf_s(outBuffer, _TRUNCATE, "%lld,0,Default,Main,0,0,0,,%s", m_iSubLineCount, str.c_str());
-                ass_process_chunk(m_track.get(), outBuffer, static_cast<int>(strnlen_s(outBuffer, sizeof(outBuffer))), tStart / 10000, (tStop - tStart) / 10000);
-            }
-        }
-    }
-}
-
-#endif
 
 bool CSimpleTextSubtitle::Open(CString provider, BYTE* data, int len, int CharSet, CString name, Subtitle::HearingImpairedType eHearingImpaired, LCID lcid)
 {
