+++ src/Subtitles/STS.h
@@ -243,36 +243,7 @@ public:
 
 public:
     STSStyle m_styleOverride; // the app can decide to use this style instead of a built-in one
-
-#if USE_LIBASS
-public:
-    bool m_renderUsingLibass;
-    OpenTypeLang::HintStr m_openTypeLangHint;
-
-    SubRendererSettings subRendererSettings;
-    void SetSubRenderSettings(SubRendererSettings settings);
-
-
-    bool m_assloaded;
-    bool m_assfontloaded;
-
-    IFilterGraph* m_pGraph;
-    std::unique_ptr<ASS_Library, ASS_LibraryDeleter> m_ass;
-    std::unique_ptr<ASS_Renderer, ASS_RendererDeleter> m_renderer;
-    std::unique_ptr<ASS_Track, ASS_TrackDeleter> m_track;
-
-    void ResetASS();
-    bool LoadASSFile(Subtitle::SubType subType);
-    bool LoadASSTrack(char* data, int size, Subtitle::SubType subType);
-    void UnloadASS();
-    void LoadASSSample(char* data, int dataSize, REFERENCE_TIME tStart, REFERENCE_TIME tStop);
-    void LoadASSFont(IPin* pPin, ASS_Library* ass, ASS_Renderer* renderer);
-    void SetFilterGraph(IFilterGraph* g) { m_pGraph = g; };
-    void SetPin(IPin* i) { m_pPin = i; };
-
-protected:
-    IPin* m_pPin;
-#endif
+    SSAUtil m_SSAUtil;
 };
 
 extern const BYTE CharSetList[];
