+++ src/mpc-hc/MainFrm.cpp
@@ -16407,14 +16407,10 @@ bool CMainFrame::LoadSubtitle(CString fn, SubtitleInput* pSubInput /*= nullptr*/
     if (!pSubStream && ext != _T(".idx") && ext != _T(".sup")) {
         CAutoPtr<CRenderedTextSubtitle> pRTS(DEBUG_NEW CRenderedTextSubtitle(&m_csSubLock));
         if (pRTS) {
-#if USE_LIBASS
             SubRendererSettings srs = AfxGetAppSettings().GetSubRendererSettings();
-            pRTS->SetSubRenderSettings(srs);
-#endif
+            pRTS->m_SSAUtil.SetSubRenderSettings(srs);
             if (pRTS->Open(fn, DEFAULT_CHARSET, _T(""), videoName) && pRTS->GetStreamCount() > 0) {
-#if USE_LIBASS
-                pRTS->SetFilterGraph(m_pGB);
-#endif
+                pRTS->m_SSAUtil.SetFilterGraph(m_pGB);
                 pSubStream = pRTS.Detach();
             }
         }
@@ -16513,7 +16509,7 @@ bool CMainFrame::LoadSubtitle(CYoutubeDLInstance::YDLSubInfo& sub) {
     if (pRTS) {
 #if USE_LIBASS
         SubRendererSettings srs = AfxGetAppSettings().GetSubRendererSettings();
-        pRTS->SetSubRenderSettings(srs);
+        pRTS->m_SSAUtil.SetSubRenderSettings(srs);
 #endif
         bool opened = false;
         if (!sub.url.IsEmpty()) {
@@ -16546,7 +16542,7 @@ bool CMainFrame::LoadSubtitle(CYoutubeDLInstance::YDLSubInfo& sub) {
         }
         if (opened && pRTS->GetStreamCount() > 0) {
 #if USE_LIBASS
-            pRTS->SetFilterGraph(m_pGB);
+            pRTS->m_SSAUtil.SetFilterGraph(m_pGB);
 #endif
             pSubStream = pRTS.Detach();
         }
@@ -20384,18 +20380,14 @@ LRESULT CMainFrame::OnLoadSubtitles(WPARAM wParam, LPARAM lParam)
 
     CAutoPtr<CRenderedTextSubtitle> pRTS(DEBUG_NEW CRenderedTextSubtitle(&m_csSubLock));
     if (pRTS) {
-#if USE_LIBASS
         SubRendererSettings srs = AfxGetAppSettings().GetSubRendererSettings();
-        pRTS->SetSubRenderSettings(srs);
-#endif
+        pRTS->m_SSAUtil.SetSubRenderSettings(srs);
         if (pRTS->Open(CString(data.pSubtitlesInfo->Provider()->Name().c_str()),
             (BYTE*)(LPCSTR)data.fileContents.c_str(), (int)data.fileContents.length(), DEFAULT_CHARSET,
             UTF8To16(data.fileName.c_str()), Subtitle::HearingImpairedType(data.pSubtitlesInfo->hearingImpaired),
             ISOLang::ISO6391ToLcid(data.pSubtitlesInfo->languageCode.c_str())) && pRTS->GetStreamCount() > 0) {
             m_wndSubtitlesDownloadDialog.DoDownloaded(*data.pSubtitlesInfo);
-#if USE_LIBASS
-            pRTS->SetFilterGraph(m_pGB);
-#endif
+            pRTS->m_SSAUtil.SetFilterGraph(m_pGB);
             SubtitleInput subElement = pRTS.Detach();
             m_pSubStreams.AddTail(subElement);
             if (data.bActivate) {
